Separate Board from the game file
