+++ packages/client/src/pages/Game.tsx
@@ -1,20 +1,19 @@
 /** @jsx jsx */
 /** @jsxRuntime classic */
 import { useEffect, useMemo, useState } from 'react';
-import { Box, Flex, Grid, Heading, jsx, Text } from 'theme-ui';
-import { Card } from 'components/Cards/Card';
+import { Box, Flex, Heading, jsx, Text } from 'theme-ui';
 import { Deck } from 'components/Cards/Deck';
 import { useUserData } from 'components/UserContext';
 import { GameTable } from 'components/GameTable';
 import { GameEvent, GameState, Score } from 'shared';
-import { cardDrop, cardWrapper, mainDeckPosition } from 'components/Cards/style';
-import { CardWrapper } from 'components/Cards/CardWrapper';
+import { mainDeckPosition } from 'components/Cards/style';
 import { sum } from 'lodash';
 import { cardKey, fromCardKey } from 'utils/cards';
 import { Opponent } from '../components/Players/Opponent';
 import { Player } from '../components/Players/Player';
 import { gameIO } from 'lib/socket';
 import { useParams } from 'react-router-dom';
+import { Board } from 'components/Board';
 
 export const Game = ({ gameState, gameScore }: { gameState: GameState; gameScore?: Score[] }) => {
   const [activePlayerCard, togglePlayerActiveCard] = useState<string | null>(null);
@@ -31,6 +30,7 @@ export const Game = ({ gameState, gameScore }: { gameState: GameState; gameScore
   console.log(gameState);
   useEffect(() => {
     let timer: NodeJS.Timeout;
+
     if (activePlayerCard && activeCardsOnTable) {
       const { value: playerCardNumber } = fromCardKey(activePlayerCard);
       const tableCardsSum = sum(activeCardsOnTable.map((c) => fromCardKey(c).value));
@@ -90,28 +90,14 @@ export const Game = ({ gameState, gameScore }: { gameState: GameState; gameScore
     <GameTable>
       <Deck cardNumber={deck.length} title={`${deck.length} cards left`} sx={mainDeckPosition} />
       <Opponent player={opponent} sx={{ transform: 'rotate(180deg)' }} />
-      <Grid sx={{ mx: 3, alignContent: 'center', flex: 1 }} columns="1fr 1fr 1fr 1fr" gap={3}>
-        {table.map((c) => {
-          const key = cardKey(c);
-          const isActive = activeCardsOnTable.includes(key);
-          const needsToMove = movingCards.includes(key);
-          return (
-            <CardWrapper
-              key={key}
-              isMoving={needsToMove}
-              sx={cardWrapper(isActive)}
-              onClick={() => {
-                toggleActiveCardsOnTable(
-                  isActive ? activeCardsOnTable.filter((c) => c !== key) : [...activeCardsOnTable, key],
-                );
-              }}
-            >
-              <Card card={c} />
-            </CardWrapper>
-          );
-        })}
-        {activePlayerCard && <Box role="button" sx={cardDrop} onClick={playCardOnTable} />}
-      </Grid>
+      <Board
+        table={table}
+        activeCardsOnTable={activeCardsOnTable}
+        movingCards={movingCards}
+        toggleActiveCardsOnTable={toggleActiveCardsOnTable}
+        activePlayerCard={activePlayerCard}
+        playCardOnTable={playCardOnTable}
+      />
       {gameScore && (
         <Flex sx={{ gap: 3 }}>
           {gameScore.map(({ details, total }, i) => (
